backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
requests==2.31.0
python-dotenv==1.0.0
chromadb==0.5.0
sentence-transformers==2.7.0
pytest==7.4.0
pytest-mock==3.12.0
gunicorn==21.2.0
